# Performance: Limit Renders in the Layers Panel