--- a/packages/story-editor/src/app/story/index.js
+++ b/packages/story-editor/src/app/story/index.js
@@ -13,13 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
+/**
+ * Internal dependencies
+ */
+import * as storyReducers from './useStoryReducer/reducers';
 export { default as StoryProvider } from './storyProvider';
 export { default as StoryContext } from './context';
-export { default as useStory } from './useStory';
+export { default as useStory, useCurrentPageElementIds } from './useStory';
 export {
   useStoryTriggers,
   useStoryTriggerListener,
   useStoryTriggersDispatch,
   STORY_EVENTS,
 } from './storyTriggers';
+
+export { storyReducers };