--- a/packages/story-editor/src/app/story/index.js
+++ b/packages/story-editor/src/app/story/index.js
@@ -19,7 +19,7 @@
 import * as storyReducers from './useStoryReducer/reducers';
 export { default as StoryProvider } from './storyProvider';
 export { default as StoryContext } from './context';
-export { default as useStory, useCurrentPageElementIds } from './useStory';
+export { default as useStory } from './useStory';
 export {
   useStoryTriggers,
   useStoryTriggerListener,