--- a/packages/story-editor/src/app/story/useStory.js
+++ b/packages/story-editor/src/app/story/useStory.js
@@ -35,6 +35,10 @@ const DELIMITER = ':::';
  * @return {Array} array of current pages story element ids in order
  */
 export function useCurrentPageElementIds() {
+  // Joining and splitting with delimiter so that selector returns
+  // a shallow equal result when we have a new id array instance that
+  // is deeply equal. This prevents empty forced re-renders from
+  // useContextSelector
   const elementIds = useStory((value) => {
     const elements = value.state.currentPage?.elements || [];
     return elements.map((element) => element.id).join(DELIMITER);