--- a/packages/story-editor/src/app/story/useStory.js
+++ b/packages/story-editor/src/app/story/useStory.js
@@ -27,4 +27,19 @@ function useStory(selector) {
   return useContextSelector(Context, selector ?? identity);
 }
 
+const DELIMITER = ':::';
+/**
+ * Returns an array of current page's story element ids. Only rerenders when there are
+ * element reorders, deletions or additions
+ *
+ * @return {Array} array of current pages story element ids in order
+ */
+export function useCurrentPageElementIds() {
+  const elementIds = useStory((v) => {
+    const elements = v.state.currentPage?.elements || [];
+    return elements.map((element) => element.id).join(DELIMITER);
+  });
+  return elementIds.split(DELIMITER);
+}
+
 export default useStory;