--- a/packages/story-editor/src/app/story/useStoryReducer/reducer.js
+++ b/packages/story-editor/src/app/story/useStoryReducer/reducer.js
@@ -91,6 +91,10 @@ function reducer(state, { type, payload }) {
       return reducers.toggleElement(state, payload);
     }
 
+    case types.TOGGLE_LAYER: {
+      return reducers.toggleLayer(state, payload);
+    }
+
     case types.DUPLICATE_ELEMENT_BY_ID: {
       return reducers.duplicateElementById(state, payload);
     }