--- a/packages/story-editor/src/components/panels/design/layer/layer.js
+++ b/packages/story-editor/src/components/panels/design/layer/layer.js
@@ -27,7 +27,7 @@ import {
   themeHelpers,
   Tooltip,
 } from '@web-stories-wp/design-system';
-import { useRef } from '@web-stories-wp/react';
+import { useRef, memo } from '@web-stories-wp/react';
 
 /**
  * Internal dependencies
@@ -252,31 +252,25 @@ function preventReorder(e) {
   e.preventDefault();
 }
 
-function Layer({ layer }) {
-  const { LayerIcon, LayerContent } = getDefinitionForType(layer.type);
-  const { isSelected, handleClick } = useLayerSelection(layer);
-  const { currentPage, deleteElementById, duplicateElementById } = useStory(
-    (state) => ({
-      currentPage: state.state.currentPage,
-      deleteElementById: state.actions.deleteElementById,
-      duplicateElementById: state.actions.duplicateElementById,
-    })
-  );
+function Layer({ element }) {
+  const { LayerIcon, LayerContent } = getDefinitionForType(element.type);
+  const { isSelected, handleClick } = useLayerSelection(element);
+  const duplicateElementById = useStory((v) => v.actions.duplicateElementById);
+  const deleteElementById = useStory((v) => v.actions.deleteElementById);
 
   const layerRef = useRef(null);
   usePerformanceTracking({
     node: layerRef.current,
-    eventData: { ...TRACKING_EVENTS.SELECT_ELEMENT, label: layer.type },
+    eventData: { ...TRACKING_EVENTS.SELECT_ELEMENT, label: element.type },
   });
 
   const deleteButtonRef = useRef(null);
   usePerformanceTracking({
     node: deleteButtonRef.current,
-    eventData: { ...TRACKING_EVENTS.DELETE_ELEMENT, label: layer.type },
+    eventData: { ...TRACKING_EVENTS.DELETE_ELEMENT, label: element.type },
   });
 
-  const isBackground = currentPage.elements[0].id === layer.id;
-  const layerId = `layer-${layer.id}`;
+  const layerId = `layer-${element.id}`;
 
   return (
     <LayerContainer>
@@ -287,17 +281,17 @@ function Layer({ layer }) {
         isSelected={isSelected}
       >
         <LayerIconWrapper>
-          <LayerIcon element={layer} currentPage={currentPage} />
+          <LayerIcon element={element} />
         </LayerIconWrapper>
         <LayerDescription>
           <LayerContentContainer>
-            {isBackground ? (
+            {element.isBackground ? (
               <LayerText>{__('Background', 'web-stories')}</LayerText>
             ) : (
-              <LayerContent element={layer} />
+              <LayerContent element={element} />
             )}
           </LayerContentContainer>
-          {isBackground && (
+          {element.isBackground && (
             <IconWrapper>
               <Icons.LockClosed />
             </IconWrapper>
@@ -305,7 +299,7 @@ function Layer({ layer }) {
         </LayerDescription>
       </LayerButton>
       <ActionsContainer>
-        {isBackground ? (
+        {element.isBackground ? (
           <LayerAction
             aria-label={__('Locked', 'web-stories')}
             aria-describedby={layerId}
@@ -325,7 +319,7 @@ function Layer({ layer }) {
                 aria-label={__('Delete', 'web-stories')}
                 aria-describedby={layerId}
                 onPointerDown={preventReorder}
-                onClick={() => deleteElementById({ elementId: layer.id })}
+                onClick={() => deleteElementById({ elementId: element.id })}
               >
                 <Icons.Trash />
               </LayerAction>
@@ -339,7 +333,7 @@ function Layer({ layer }) {
                 aria-label={__('Duplicate', 'web-stories')}
                 aria-describedby={layerId}
                 onPointerDown={preventReorder}
-                onClick={() => duplicateElementById({ elementId: layer.id })}
+                onClick={() => duplicateElementById({ elementId: element.id })}
               >
                 <Icons.PagePlus />
               </LayerAction>
@@ -352,7 +346,7 @@ function Layer({ layer }) {
 }
 
 Layer.propTypes = {
-  layer: StoryPropTypes.layer.isRequired,
+  element: StoryPropTypes.element.isRequired,
 };
 
-export default Layer;
+export default memo(Layer);