--- a/packages/story-editor/src/components/panels/design/layer/layer.js
+++ b/packages/story-editor/src/components/panels/design/layer/layer.js
@@ -142,6 +142,37 @@ const LayerButton = styled(Button).attrs({
     --background-color-opaque: ${({ theme }) =>
       rgba(theme.colors.interactiveBg.secondaryPress, 0)};
   }
+
+      background: ${theme.colors.interactiveBg.secondaryPress};
+      + * {
+        --background-color: ${theme.colors.interactiveBg.secondaryPress};
+        --background-color-opaque: ${rgba(
+          theme.colors.interactiveBg.secondaryPress,
+          0
+        )};
+        --selected-hover-color: ${theme.colors.interactiveBg.tertiaryHover};
+      }
+    `}
+
+  :hover {
+    background: ${({ theme }) => theme.colors.interactiveBg.secondaryHover};
+  }
+  :hover + * {
+    --background-color: ${({ theme }) =>
+      theme.colors.interactiveBg.secondaryHover};
+    --background-color-opaque: ${({ theme }) =>
+      rgba(theme.colors.interactiveBg.secondaryHover, 0)};
+  }
+
+  :active {
+    background: ${({ theme }) => theme.colors.interactiveBg.secondaryPress};
+  }
+  :active + * {
+    --background-color: ${({ theme }) =>
+      theme.colors.interactiveBg.secondaryPress};
+    --background-color-opaque: ${({ theme }) =>
+      rgba(theme.colors.interactiveBg.secondaryPress, 0)};
+  }
 `;
 
 const LayerIconWrapper = styled.div`
@@ -255,11 +286,11 @@ function preventReorder(e) {
 function Layer({ element }) {
   const { LayerIcon, LayerContent } = getDefinitionForType(element.type);
   const { isSelected, handleClick } = useLayerSelection(element);
-  const duplicateElementById = useStory(
-    (value) => value.actions.duplicateElementById
-  );
-  const deleteElementById = useStory(
-    (value) => value.actions.deleteElementById
+  const { duplicateElementById, deleteElementById } = useStory(
+    ({ actions }) => ({
+      duplicateElementById: actions.duplicateElementById,
+      deleteElementById: actions.deleteElementById,
+    })
   );
 
   const layerRef = useRef(null);