--- a/packages/story-editor/src/components/panels/design/layer/layerPanel.js
+++ b/packages/story-editor/src/components/panels/design/layer/layerPanel.js
@@ -54,7 +54,7 @@ function LayerPanel() {
   const layers = useLayers();
 
   const inspectorContentHeight = useInspector(
-    (value) => value.state.inspectorContentHeight
+    ({ state }) => state.inspectorContentHeight
   );
 
   // We want the max height to fill the space underneath the document/design tab bar.