## Context
Performance optimization & documentation around several key interactions in the Layers Panel. Makes key actions only render relevant components in the Layers Panel. Basically now `O(updated layers)` vs `O(all layers)` for these interactions

## Summary
Improving rendering performance of the layers panel with 3 interactions. 

Tested on a page with 50 square shape elements (51 including the background element) with the Layers panel open:
<img width="1440" alt="Screen Shot 2021-12-22 at 1 55 36 PM" src="https://user-images.githubusercontent.com/35983235/147141603-77df3d34-db93-4ac2-bea7-bcc38644f492.png">

*Tested on a fully specced 2018 Macbook Pro with no CPU Throttling*


### Element Selection
Test: Changing the selected element in the editor by clicking a different layer in the layers panel. 
Result: **~49.14x faster** 

| Before | After |
| --- | --- |
| ~68.8ms | ~1.4ms |
|  <img width="648" alt="Screen Shot 2021-12-22 at 1 28 19 PM" src="https://user-images.githubusercontent.com/35983235/147138712-dbc7872c-15cb-4d97-b404-41ed0cd51a5d.png"> |  <img width="684" alt="Screen Shot 2021-12-22 at 1 28 29 PM" src="https://user-images.githubusercontent.com/35983235/147138731-9ba24179-7f32-4998-9885-642f42c4c01c.png"> |


### Reordering Layers
Test: Drag the selected layer in the layers panel to a new position. 
Result: **~45.48x faster** 

| Before | After |
| --- | --- |
| ~104.6ms | ~2.3ms |
| <img width="600" alt="Screen Shot 2021-12-22 at 1 42 27 PM" src="https://user-images.githubusercontent.com/35983235/147141054-88360c27-e4cf-46bb-ba38-f223036d1cfa.png"> |  <img width="567" alt="Screen Shot 2021-12-22 at 1 42 38 PM" src="https://user-images.githubusercontent.com/35983235/147141071-62a10d1c-0790-4dd3-a03c-829f992bb2ea.png"> |



### Updating an elements style properties
Test: Drag the previously selected element on the canvas to update its x & y coordinates
Result: **~42.75x faster** 

| Before | After |
| --- | --- |
| <img width="599" alt="Screen Shot 2021-12-22 at 2 01 56 PM" src="https://user-images.githubusercontent.com/35983235/147142358-c0b27c89-896b-4917-ae74-54e5e7e85fc8.png"> |  <img width="600" alt="Screen Shot 2021-12-22 at 2 13 26 PM" src="https://user-images.githubusercontent.com/35983235/147143682-ad3d5d85-71ac-4bd5-8386-5544bffcf31b.png"> |




## Relevant Technical Choices
Introduces better (more sensitive to referential equality) patterns around `useContextSelector` more inline with the library's recommended usage:
https://github.com/dai-shi/use-context-selector#usage

<!-- Please describe your changes. -->

## To-do

<!-- A list of things that need to be addressed in this PR or follow-up changes. -->

## User-facing changes

<!--
Please describe your changes.
Include before/after screenshots or a short video.
-->

## Testing Instructions

<!--
How can the changes in this PR be verified?
Please provide step-by-step instructions how to reproduce the issue, if applicable.
Write step-by-step test instructions aimed at non-tech-savvy users, even if the PR is not user-facing.
-->

<!-- ignore-task-list-start -->
- [x] This is a non-user-facing change and requires no QA
<!-- ignore-task-list-end -->

This PR can be tested by following these steps:

1.


## Reviews

### Does this PR have a security-related impact?

<!-- Examples: new APIs, changes to KSES, etc.  -->

### Does this PR change what data or activity we track or use?

<!-- Examples: changes to telemetry, new third-party APIs -->

### Does this PR have a legal-related impact?

<!-- Examples: new images with unknown sources, new production dependencies with incompatible licenses -->

## Checklist

<!-- Check these after PR creation -->

- [x] This PR addresses an existing issue and I have linked this PR to it in ZenHub
- [x] I have tested this code to the best of my abilities
- [x] I have verified accessibility to the best of my abilities ([docs](https://github.com/google/web-stories-wp/blob/main/docs/accessibility-testing.md))
- [x] I have verified i18n and l10n (translation, right-to-left layout) to the best of my abilities
- [x] This code is covered by automated tests (unit, integration, and/or e2e) to verify it works as intended ([docs](https://github.com/google/web-stories-wp/tree/main/docs#testing))
- [x] I have added documentation where necessary
- [x] I have added a matching `Type: XYZ` label to the PR

---

<!--
Please reference the issue(s) this PR addresses.
No URLs, just the issue numbers.
Use "Fixes #123" if it fixes an issue.

NOTE: One reference per line!

Example:

Fixes #123
Partially addresses #456
See #789
-->

Fixes #10051