@littlemilkstudio 2021-12-24
@swissspidy @choumx If y'all are at all interested, I'd be happy to do a little meeting over the thought process behind our `use-context-selector` patterns introduced here.

Writing about it can get a little verbose, but it's relatively simple concepts.

Happy to write up a reference doc too if that's preferred, or if you'd like it in addition to a little 20-30 minute meeting.

Let me know. Happy Holidays 🥂